UI fixes and post-fixing
There were some bugs in the UI I fixed, I changed the colours to be ever so slightly less blinding, and I fixed some of the formatting of the preview to my post so it's only the first paragraph now.

New post soon probably if I do it after I take a shower.